requirements.txt
python-Levenshtein==0.26.1
regex==2024.11.6
numpy==2.2.3
torch==2.5.1
transformers==4.49.0
datasets==3.3.1
matplotlib==3.10.0
accelerate==1.4.0
scipy==1.14.1
seaborn==0.13.2
textstat==0.7.10
scikit-learn==1.7.1
pytest==8.4.2
pandas==2.2.3
sentencepiece==0.2.0